feat: multi cluster support
Other changes: missing dns entries in /etc/hosts. Application now is able to create cluster using user other than root. Add cluster name, extract cluster settings to separated step in wizard, remote project stored in Proxmox

+++ internal/utils/retry.go
@@ -2,9 +2,9 @@ package utils
 
 import "time"
 
-func Retry(attempts int, waitTime time.Duration, task func(attempt int) error) error {
+func Retry(attempts uint, waitTime time.Duration, task func(attempt uint) error) error {
 	var lastError error = nil
-	for i := 1; i <= attempts+1; i++ {
+	for i := uint(1); i <= attempts+1; i++ {
 		if i == attempts+1 {
 			return lastError
 		}
